# lib/api-zod/src/generated/types/busDeparture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusDeparture {
  /** ISO 8601 datetime string */
  expectedArrivalTime: string;
  /** Minutes until arrival (can be negative if past) */
  minutesUntilArrival: number;
  /** Final destination name */
  destination: string;
  /** Arrival time formatted as HH:MM */
  formattedTime: string;
  /** Vehicle reference ID */
  vehicleRef: string;
}