# lib/api-zod/src/generated/types/nextBusesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusDeparture } from "./busDeparture";

export interface NextBusesResponse {
  departures: BusDeparture[];
  /** Name of the stop */
  stopName: string;
  /** ISO 8601 datetime string of when data was fetched */
  lastUpdated: string;
}